requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
cryptography==3.4.7
docopt==0.6.2
idna==3.2
mysql-connector-python==8.0.26
pipreqs==0.4.10
protobuf==3.17.3
pycparser==2.20
requests==2.26.0
six==1.16.0
tabulate==0.8.9
urllib3==1.26.6
yarg==0.1.9
